Add support for sass/scss @import
I found out the hard way that Scalate's support for sass @import was
broken. It seems it didn't know where to look (in the filesystem) for
any imported templates.

This change populates the `load_paths` sass exec option with the
`TemplateEngine`'s `sourceDirectories` list, letting sass know where to
look for files it wants to import.
I think this change makes the sass module behave the same as other
Scalate modules with respect to path searching.