[#120] Made some classloader changes for running in an OSGi environment.
We use the same classloader selection mechanism at runtime as was there
for compile-time choosing for the OSGi/non OSGi environments. This
enables the running (compiled) template to reference other classes in
the classpath, but not necessarily in the classloader of
getClass().getClassLoader().

Also fixed the initialization code of the ScalaCompiler and
OsgiScalaCompiler to use a method for compiler creation. The
initialization of the ScalaCompiler.compiler was happening even when the
subclass overrode the .compiler field, causing it to fail in an OSGi
environment.  Now we just override the creation method, and it works as
expected.