Reset scala Source before read to support multiple reads
A Scala Source can only be read once and then has to be reset in order to
be read again. Therefore, in SourceResource, if the position is not 0,
close the existing resource, and then call reset to ensure the Source can
be read from the beginning.

Some Scala Source implementations provide reset out of the box, however
others (such as ones created by Source.fromInputStream) need the creater
of the Source to provide the reset function via the withReset method.