Avoid duplicate TemplateSource reads unless using pipeline
TemplateEngine was calling source.text even if pipelines were not being
used, causing the text function of the underlying Resource to be called
more than once.

Move the call to source.text into the match for pipeline so that the
underlying Resource is read only once unless pipeline's are being used,
in which case the Resource will (currently) be read twice as before.